+++ src/App.js
@@ -2,11 +2,11 @@ import React from 'react';
 import {BrowserRouter as Router, Route, Link} from "react-router-dom";
 import 'bootstrap/dist/css/bootstrap.css';
 
-import LandingPage from "./scenes/LandingPage";
+import LandingPage from "./scenes/LandingPage/LandingPage";
 import Navbar from "./components/Navbar";
 import Footer from "./components/Footer";
-import Blog from "./scenes/Blog";
-import Scholarship from "./scenes/Scholarship";
+import Blog from "./scenes/Blog/Blog";
+import Scholarship from "./scenes/Scholarship/Scholarship";
 
 
 function App() {
