+++ src/scenes/LandingPage/LandingPage.scss
@@ -1,4 +1,4 @@
-@import '../index.scss';
+@import '../../index';
 
 .background-image {
   //background: linear-gradient(to left, #00c6ff, #30a0ff);
